Refactor lib into dedicated module files
